Removed unnecessary $this-> in the BookEvent class and added a search.php page
I realized I was repeating a lot of things using $this-> and creating properties I wasn't using (or using in an ineficient way). So I decided to remove them and make sure methods didn't use properties of the class. I'm still learning this and I hope to get better at it in the future.